+++ src/promptflow-evals/promptflow/evals/evaluators/_protected_materials/__init__.py
@@ -1,0 +1,5 @@
+from ._protected_materials import ProtectedMaterialsEvaluator
+
+__all__ = [
+    "ProtectedMaterialsEvaluator",
+]